refactor(engine): extract the deny-hook pure logic into gittensory-engine
lib/deny-hooks.js's rule evaluator and lib/deny-hook-synthesis.js's
proposal-synthesis logic were pure, deterministic business logic living
directly in the self-host gittensory-miner package with no gittensory-engine
counterpart -- unlike every other business-logic module (governor/chokepoint,
portfolio-queue selection, worktree planning, ...), which already follow the
"pure engine function + thin miner-lib wrapper" split. A hosted service wanting
per-tenant deny-hook enforcement had to depend on the whole SQLite-backed miner
package just to reach one pure function.

Mirroring governor/chokepoint.ts's split:
- New packages/gittensory-engine/src/miner/deny-hooks.ts holds the pure rule
  evaluator (DEFAULT_DENY_RULES, evaluateDenyHooks + types).
- New packages/gittensory-engine/src/miner/deny-hook-synthesis.ts holds the pure
  proposal-synthesis logic (synthesizeDenyRuleProposals, aggregateBlockerHistory,
  resolveEffectiveDenyRules, setProposalStatuses, the normalize/canonicalize
  helpers, DEFAULT_SYNTHESIS_CONFIG + types).
- The audit-stamp clock is now injected: the engine's buildPathProposal /
  synthesizeDenyRuleProposals take an explicit nowMs, so the extracted logic is
  fully deterministic given its inputs (no direct new Date()).
- Both are re-exported through @loopover/engine's public barrel.
- lib/deny-hooks.js is now a thin re-export; lib/deny-hook-synthesis.js keeps
  ONLY the SQLite store (initDenyHookSynthesisStore etc.) and a 2-arg
  synthesizeDenyRuleProposals wrapper that defaults nowMs to Date.now(), so its
  public API and behavior are unchanged.

No behavior change: the four existing deny-hook/synthesis suites pass unchanged,
plus a new regression test proving the extracted functions are callable from
@loopover/engine alone (no miner/SQLite dependency) and that the injected clock
makes synthesizedAt deterministic.

Closes #5667
