+++ packages/gittensory-miner/lib/deny-hook-synthesis.js
@@ -1,30 +1,53 @@
-// Synthesize PreToolUse deny-hook rule proposals from per-repo blocker/path history (#4522). Pure synthesis
-// plus an optional local SQLite store for refresh + maintainer review before any synthesized rule takes effect.
-// Approved rules merge with {@link DEFAULT_DENY_RULES}; unapproved proposals never block tool calls. Feeds the
-// consumption surface #2343 will wire into evaluateDenyHooks — this issue owns derivation + audit, not live hook
-// interception.
+// Synthesize PreToolUse deny-hook rule proposals from per-repo blocker/path history (#4522). The pure synthesis
+// logic moved into `@loopover/engine` (packages/gittensory-engine/src/miner/deny-hook-synthesis.ts) by #5667;
+// this module is now a thin wrapper that re-exports those pure helpers and keeps the local SQLite store for
+// refresh + maintainer review before any synthesized rule takes effect. Approved rules merge with
+// {@link DEFAULT_DENY_RULES}; unapproved proposals never block tool calls. No behavior change.
 import { chmodSync, mkdirSync } from "node:fs";
 import { homedir } from "node:os";
 import { dirname, join } from "node:path";
-import { createHash } from "node:crypto";
 import { DatabaseSync } from "node:sqlite";
-import { DEFAULT_DENY_RULES, evaluateDenyHooks } from "./deny-hooks.js";
+import {
+  aggregateBlockerHistory,
+  canonicalizeChangedPath,
+  changedPathToDenyGlob,
+  DEFAULT_SYNTHESIS_CONFIG,
+  isCoveredByDefaultDenyRules,
+  normalizeBlockerHistory,
+  normalizeBlockerHistoryRecord,
+  normalizeRepoFullName,
+  proposalStatusSet,
+  resolveEffectiveDenyRules,
+  setProposalStatuses,
+  synthesizeDenyRuleProposals as engineSynthesizeDenyRuleProposals,
+} from "@loopover/engine";
 import { DEFAULT_FORGE_CONFIG } from "./forge-config.js";
 
+// Re-export the pure synthesis helpers from the engine so this module's public API is unchanged after #5667
+// moved derivation/audit into @loopover/engine. Only the SQLite store below (and its forge/db-path helpers) is
+// miner-local, because it depends on node:sqlite/node:fs and this package's forge-config default.
+export {
+  aggregateBlockerHistory,
+  canonicalizeChangedPath,
+  changedPathToDenyGlob,
+  DEFAULT_SYNTHESIS_CONFIG,
+  isCoveredByDefaultDenyRules,
+  normalizeBlockerHistory,
+  normalizeBlockerHistoryRecord,
+  resolveEffectiveDenyRules,
+  setProposalStatuses,
+};
+
 const defaultDbFileName = "deny-hook-synthesis.sqlite3";
-const PROPOSAL_STATUSES = Object.freeze(["proposed", "approved", "rejected"]);
-const proposalStatusSet = new Set(PROPOSAL_STATUSES);
 
-export const DEFAULT_SYNTHESIS_CONFIG = Object.freeze({
-  minPathOccurrences: 2,
-  maxProposals: 20,
-});
-
-function normalizeRepoFullName(repoFullName) {
-  if (typeof repoFullName !== "string") throw new Error("invalid_repo_full_name");
-  const [owner, repo, extra] = repoFullName.trim().split("/");
-  if (!owner || !repo || extra !== undefined) throw new Error("invalid_repo_full_name");
-  return `${owner}/${repo}`;
+/**
+ * Derive candidate deny-hook rules from blocker/path history. Miner-facing wrapper over the engine's pure
+ * `synthesizeDenyRuleProposals`, defaulting the injected clock to `Date.now()` so this keeps the pre-#5667 2-arg
+ * signature (and wall-clock `audit.synthesizedAt`) every existing caller and test relies on. Returns proposal
+ * objects only — nothing is active until a maintainer approves them (see resolveEffectiveDenyRules).
+ */
+export function synthesizeDenyRuleProposals(records, config = {}) {
+  return engineSynthesizeDenyRuleProposals(records, config, Date.now());
 }
 
 /** Optional forge host, scoping rows so two hosts serving the same owner/repo name never collide (#5563).
@@ -33,203 +56,6 @@ function normalizeApiBaseUrl(apiBaseUrl) {
   if (apiBaseUrl === undefined || apiBaseUrl === null) return DEFAULT_FORGE_CONFIG.apiBaseUrl;
   if (typeof apiBaseUrl !== "string" || !apiBaseUrl.trim()) throw new Error("invalid_api_base_url");
   return apiBaseUrl.trim();
-}
-
-function normalizeOptionalStringArray(value) {
-  if (value === undefined || value === null) return [];
-  if (!Array.isArray(value)) return [];
-  return value.filter((entry) => typeof entry === "string" && entry.trim()).map((entry) => entry.trim());
-}
-
-/** Validate one blocker-history row from the review stack (gate block/close audit). */
-export function normalizeBlockerHistoryRecord(record) {
-  if (!record || typeof record !== "object" || Array.isArray(record)) return null;
-  const blockerCodes = normalizeOptionalStringArray(record.blockerCodes);
-  if (blockerCodes.length === 0) return null;
-  const changedPaths = normalizeOptionalStringArray(record.changedPaths);
-  const guardrailMatches = normalizeOptionalStringArray(record.guardrailMatches);
-  const repoFullName = typeof record.repoFullName === "string" && record.repoFullName.trim()
-    ? normalizeRepoFullName(record.repoFullName)
-    : null;
-  return {
-    repoFullName,
-    blockerCodes,
-    changedPaths,
-    guardrailMatches,
-    pullNumber: Number.isInteger(record.pullNumber) && record.pullNumber > 0 ? record.pullNumber : null,
-    recordedAt: typeof record.recordedAt === "string" && record.recordedAt.trim() ? record.recordedAt.trim() : null,
-  };
-}
-
-export function normalizeBlockerHistory(records) {
-  if (!Array.isArray(records)) return [];
-  const normalized = [];
-  for (const record of records) {
-    const entry = normalizeBlockerHistoryRecord(record);
-    if (entry) normalized.push(entry);
-  }
-  return normalized;
-}
-
-/** Canonicalize a changed path the same way guardrail matching does (case/separator insensitive). */
-export function canonicalizeChangedPath(path) {
-  if (typeof path !== "string") return null;
-  const trimmed = path.trim().replace(/\\/g, "/").replace(/^\.\/+/, "").replace(/^\/+/, "");
-  if (!trimmed || trimmed.includes("..")) return null;
-  return trimmed.toLowerCase();
-}
-
-/** Convert a repo-relative changed path into a deny-hook glob matching DEFAULT_DENY_RULES shape. */
-export function changedPathToDenyGlob(path) {
-  const canonical = canonicalizeChangedPath(path);
-  if (!canonical) return null;
-  return `**/${canonical}`;
-}
-
-function ruleSignature(rule) {
-  return JSON.stringify({
-    matcher: rule.matcher,
-    pathPattern: rule.pathPattern ?? null,
-    inputIncludesAll: rule.inputIncludesAll ?? null,
-    reason: rule.reason,
-  });
-}
-
-/** True when a synthesized glob is already enforced by a built-in default deny rule. */
-export function isCoveredByDefaultDenyRules(pathPattern) {
-  if (typeof pathPattern !== "string" || !pathPattern.trim()) return false;
-  const samplePath = pathPattern.replace(/^\*\*\//, "");
-  if (!samplePath) return false;
-  return !evaluateDenyHooks({ name: "Write", input: { file_path: samplePath } }, DEFAULT_DENY_RULES).allowed;
-}
-
-function collectPathsFromRecord(record) {
-  const paths = new Set();
-  for (const path of [...record.changedPaths, ...record.guardrailMatches]) {
-    const canonical = canonicalizeChangedPath(path);
-    if (canonical) paths.add(canonical);
-  }
-  return paths;
-}
-
-/** Aggregate path and blocker-code frequencies from normalized history. Pure. */
-export function aggregateBlockerHistory(records) {
-  const normalized = normalizeBlockerHistory(records);
-  const pathCounts = new Map();
-  const pathBlockers = new Map();
-  const blockerCounts = new Map();
-
-  for (const record of normalized) {
-    for (const code of record.blockerCodes) {
-      blockerCounts.set(code, (blockerCounts.get(code) ?? 0) + 1);
-    }
-    for (const path of collectPathsFromRecord(record)) {
-      pathCounts.set(path, (pathCounts.get(path) ?? 0) + 1);
-      const blockers = pathBlockers.get(path) ?? new Set();
-      for (const code of record.blockerCodes) blockers.add(code);
-      pathBlockers.set(path, blockers);
-    }
-  }
-
-  return {
-    pathCounts,
-    pathBlockers,
-    blockerCounts,
-    recordCount: normalized.length,
-  };
-}
-
-function stableProposalId(kind, key) {
-  const digest = createHash("sha256").update(`${kind}:${key}`).digest("hex").slice(0, 16);
-  return `${kind}:${digest}`;
-}
-
-function buildPathProposal(path, occurrenceCount, blockerCodes) {
-  const pathPattern = changedPathToDenyGlob(path);
-  if (!pathPattern || isCoveredByDefaultDenyRules(pathPattern)) return null;
-  const sortedBlockers = [...blockerCodes].sort();
-  const reason = `Synthesized deny rule: ${occurrenceCount} gate block(s) touched ${path} (${sortedBlockers.join(", ") || "path history"}). Review before enabling.`;
-  const rule = { matcher: "*", pathPattern, reason };
-  return {
-    id: stableProposalId("path", pathPattern),
-    status: "proposed",
-    rule,
-    audit: {
-      kind: "path_history",
-      path,
-      pathPattern,
-      occurrenceCount,
-      blockerCodes: sortedBlockers,
-      synthesizedAt: new Date(0).toISOString(),
-    },
-  };
-}
-
-/**
- * Derive candidate deny-hook rules from blocker/path history. Returns proposal objects only — nothing is active
- * until a maintainer approves them (see resolveEffectiveDenyRules).
- */
-export function synthesizeDenyRuleProposals(records, config = {}) {
-  const minPathOccurrences = Number.isInteger(config.minPathOccurrences)
-    ? Math.max(1, config.minPathOccurrences)
-    : DEFAULT_SYNTHESIS_CONFIG.minPathOccurrences;
-  const maxProposals = Number.isInteger(config.maxProposals)
-    ? Math.max(1, config.maxProposals)
-    : DEFAULT_SYNTHESIS_CONFIG.maxProposals;
-
-  const { pathCounts, pathBlockers, recordCount } = aggregateBlockerHistory(records);
-  if (recordCount === 0) return [];
-
-  const rankedPaths = [...pathCounts.entries()]
-    .filter(([, count]) => count >= minPathOccurrences)
-    .sort((left, right) => right[1] - left[1] || left[0].localeCompare(right[0]));
-
-  const proposals = [];
-  const seenSignatures = new Set(DEFAULT_DENY_RULES.map(ruleSignature));
-  for (const [path, count] of rankedPaths) {
-    const proposal = buildPathProposal(path, count, pathBlockers.get(path) ?? new Set());
-    if (!proposal) continue;
-    const signature = ruleSignature(proposal.rule);
-    if (seenSignatures.has(signature)) continue;
-    seenSignatures.add(signature);
-    proposals.push({
-      ...proposal,
-      audit: { ...proposal.audit, synthesizedAt: new Date().toISOString() },
-    });
-    if (proposals.length >= maxProposals) break;
-  }
-  return proposals;
-}
-
-/** Merge built-in defaults with maintainer-approved synthesized rules (deduped, defaults first). */
-export function resolveEffectiveDenyRules(options = {}) {
-  const includeDefaults = options.includeDefaults !== false;
-  const approvedProposals = Array.isArray(options.approvedProposals) ? options.approvedProposals : [];
-  const merged = includeDefaults ? [...DEFAULT_DENY_RULES] : [];
-  const seen = new Set(merged.map(ruleSignature));
-  for (const proposal of approvedProposals) {
-    if (proposal?.status !== "approved") continue;
-    const rule = proposal.rule;
-    if (!rule || typeof rule !== "object") continue;
-    const signature = ruleSignature(rule);
-    if (seen.has(signature)) continue;
-    seen.add(signature);
-    merged.push(rule);
-  }
-  return merged;
-}
-
-/** Apply maintainer approval/rejection to in-memory proposals. Pure. */
-export function setProposalStatuses(proposals, updates) {
-  if (!Array.isArray(proposals)) throw new Error("invalid_proposals");
-  const updateMap = updates instanceof Map
-    ? updates
-    : new Map(Object.entries(updates ?? {}).filter(([id]) => typeof id === "string"));
-  return proposals.map((proposal) => {
-    const nextStatus = updateMap.get(proposal.id);
-    if (!nextStatus || !proposalStatusSet.has(nextStatus)) return proposal;
-    return { ...proposal, status: nextStatus };
-  });
 }
 
 export function resolveDenyHookSynthesisDbPath(env = process.env) {
